radicle-surf: design document
The work on the refactoring document[0] has been extremely useful in
pinpointing problem areas in the radicle-surf redesign.

This document intends to supplement that by outlining a high-level API
that the redesign can follow. The design specifies a series of
components and recommended ways of implementing those components. The
exact implementation details are avoided. It also aims to provide more
education around different concepts in git by supplying git
documentation and explanations of the different types one may come
across in git.

[0]: https://github.com/radicle-dev/radicle-git/blob/main/radicle-surf/docs/refactor-design.md

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>