git-commit: remove dependency on git-trailers
The git-trailers crate depends on parsing logic using nom. This is a
heavy dependency for something that can be done using git2.

Remove the dependency of git-trailers in git-commit and localise the
trailer types -- using git2 to parse the trailers out of the git
commit message.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
X-Clacks-Overhead: GNU Terry Pratchett