Add line Modification for diffs
When source code changes (patches) are proposed to a project, the review
process tends to result in iterations on the patch.  It is helpful to
reviewers to show only the changes between the iterations on a patch,
the diff between each Patch revision's diff.

Add a new line modification type, similar to the `Modification` type,
which describes a line change between diffs.

Signed-off-by: Slack Coder <slackcoder@server.ky>