Re-License to "MIT OR APACHE"
The main dependant of `radicle-git`, and its crates, is [`heartwood`].
It is licensed under MIT OR APACHE, which has been incompatible with the GPL-3.0 license.
Since the intellectual property of this code base is with the Better Internet Foundation, it is
decided that this repository will be re-licensed to make the upstream repository compatible.

The changes required were:
- Remove the license headers from files – and the instructions to add them
- Add MIT and Apache licenses
- Remove the GPL-3.0 license
- Change the license declarations in the `Cargo.toml` files

[`heartwood`]: https://app.radicle.xyz/nodes/seed.radicle.xyz/rad:z3gqcJUoA1n9HaHKufZs5FCSGazv5